+++ backend/packages/sandbox/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @cat-factory/sandbox
 
+## 0.8.21
+
+### Patch Changes
+
+- Updated dependencies [bbafec9]
+- Updated dependencies [bbafec9]
+  - @cat-factory/kernel@0.43.0
+  - @cat-factory/agents@0.20.3
+  - @cat-factory/sandbox-fixtures@0.7.40
+
+## 0.8.20
+
+### Patch Changes
+
+- Updated dependencies [63e2177]
+  - @cat-factory/contracts@0.41.0
+  - @cat-factory/agents@0.20.2
+  - @cat-factory/kernel@0.42.2
+  - @cat-factory/sandbox-fixtures@0.7.40
+
+## 0.8.19
+
+### Patch Changes
+
+- Updated dependencies [d1027ec]
+  - @cat-factory/contracts@0.40.1
+  - @cat-factory/kernel@0.42.1
+  - @cat-factory/agents@0.20.1
+  - @cat-factory/sandbox-fixtures@0.7.39
+
 ## 0.8.18
 
 ### Patch Changes